refactor: user -> auth로 수정
