+++ AuthService/.gitattributes
