+++ AuthService/settings.gradle
