+++ AuthService/src/main/java/ready_to_marry/authservice/AuthServiceApplication.java
@@ -1,13 +1,13 @@
-package ready_to_marry.userservice;
+package ready_to_marry.authservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class UserServiceApplication {
+public class AuthServiceApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(UserServiceApplication.class, args);
+        SpringApplication.run(AuthServiceApplication.class, args);
     }
 
 }
