+++ AuthService/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.application.name=UserService
+spring.application.name=AuthService
 
 server.port = 8081
