+++ Jenkinsfile
@@ -1,9 +1,9 @@
 pipeline {
     agent any
     environment {
-	    SERVICE_REPO_NAME = 'RTM_UserService.git'
-	    DOCKER_IMAGE = 'hyundooboo/rtm-user-service'
-	    SERVICE_NAME = 'user-service'
+	    SERVICE_REPO_NAME = 'RTM_AuthService.git'
+	    DOCKER_IMAGE = 'hyundooboo/rtm-auth-service'
+	    SERVICE_NAME = 'auth-service'
 	    SPRING_DATASOURCE_URL = credentials('postgres-url')
 	    DB_CREDENTIALS = credentials('postgres-db-credentials')
 	    GITHUB_CREDENTIALS = credentials('github-token')
